--- app/assets/stylesheets/spree/backend/spree_reviews.css
@@ -1,3 +1,0 @@
-/*
- *= require spree/backend
-*/
