+++ app/assets/stylesheets/spree/frontend/spree_reviews.css
@@ -1,0 +1,111 @@
+.star-box {
+  background: white;
+  width: 30px;
+  height: 30px;
+  border-radius: 3px;
+  border: 1px solid lightgrey;
+  display: inline-block;
+  text-align: center !important;
+}
+
+.review-comment {
+  text-transform: uppercase;
+  font-size: 1.2em;
+}
+
+.card-header{
+	position: relative;
+  border-bottom: none;
+  background-color: rgba(0, 0, 0, 0.08);
+}
+.card-header:after {
+	content: '';
+	position: absolute;
+	bottom: 0;
+	left: 50%;
+	width: 0;
+	height: 0;
+	border: 23px solid transparent;
+	border-top-color: rgba(0, 0, 0, 0.08);
+	border-bottom: 0;
+	margin-left: -23px;
+	margin-bottom: -23px;
+}
+
+#rev{
+  max-width: 90%;
+  margin:auto;
+  padding: 15px;
+}
+
+.center-ratings {
+  margin: auto;
+  width: 190px;
+  text-align: center;
+  height: 30px;
+  display: flex;
+}
+.rate-area {
+    float:left;
+    border-style: none;
+}
+
+.rate-area:not(:checked) > input {
+    position:absolute;
+    top:-9999px;
+    clip:rect(0,0,0,0);
+}
+
+.rate-area:not(:checked) > label {
+    float:right;
+    overflow:hidden;
+    white-space:nowrap;
+    cursor:pointer;
+    font-size:100%;
+    color:lightgrey;
+    line-height: 1.8em;
+    margin: 0 2px;
+}
+
+.rate-area:not(:checked) > label:before {
+  content: '\2605';
+}
+
+.rate-area > input:checked ~ label {
+    color: white;
+    background: #3cbb91;
+    border-color: #3cbb91;
+}
+
+.rate-area:not(:checked) > label:hover,
+.rate-area:not(:checked) > label:hover ~ label {
+  color: white;
+  background: #3cbb91;
+  border-color: #3cbb91;
+}
+
+.rate-area > input:checked + label:hover,
+.rate-area > input:checked + label:hover ~ label,
+.rate-area > input:checked ~ label:hover,
+.rate-area > input:checked ~ label:hover ~ label,
+.rate-area > label:hover ~ input:checked ~ label {
+  color: white;
+  background: #3cbb91;
+  border-color: #3cbb91;
+}
+
+.display-rating {
+  pointer-events: none !important;
+}
+
+.carousel-control-prev-icon {
+ background-image: url("data:image/svg+xml;charset=utf8,%3Csvg xmlns='http://www.w3.org/2000/svg' fill='black' viewBox='0 0 8 8'%3E%3Cpath d='M5.25 0l-4 4 4 4 1.5-1.5-2.5-2.5 2.5-2.5-1.5-1.5z'/%3E%3C/svg%3E") !important;
+}
+
+.carousel-control-next-icon {
+  background-image: url("data:image/svg+xml;charset=utf8,%3Csvg xmlns='http://www.w3.org/2000/svg' fill='black' viewBox='0 0 8 8'%3E%3Cpath d='M2.75 0l-1.5 1.5 2.5 2.5-2.5 2.5 1.5 1.5 4-4-4-4z'/%3E%3C/svg%3E") !important;
+}
+
+.carousel-control-prev, .carousel-control-next {
+  width: 5%;
+}
