--- app/assets/stylesheets/spree/frontend/spree_reviews.css.erb
@@ -1,83 +1,0 @@
-/*
- *= require spree/frontend
-*/
-
-#reviews div.header {
-  margin-top: 10px;
-  margin-bottom: 10px;
-}
-
-.lit {
-  background-color: #aaaaaa;
-  color: #0000bb;
-  margin-right: 1px;
-  padding-left: 1px;
-}
-
-.unlit {
-  background-color: #aaaaaa;
-  color: #000000;
-  margin-right: 1px;
-  padding-left: 1px;
-}
-
-.star-rating-control {
-  display: inline-block;
-  height: 14px;
-}
-
-.review_hr {
-  size: 1px;
-  background: silver;
-  border: none;
-  margin: 1em 0;
-}
-
-/* jQuery.Rating Plugin CSS - http://www.fyneworks.com/jquery/star-rating/ */
-div.rating-cancel, div.star-rating {
-  float: left;
-  width: 17px;
-  height: 16px;
-  text-indent: -999em;
-  cursor: pointer;
-  display: block;
-  background: transparent;
-  overflow: hidden
-}
-
-div.rating-cancel, div.rating-cancel a {
-  background: transparent url(<%= asset_path('store/reviews/delete.gif') %>) no-repeat scroll 0 -1px;
-}
-
-div.star-rating, div.star-rating a {
-  background: url(<%= asset_path('store/reviews/star.gif') %>) no-repeat 0 0px
-}
-
-div.rating-cancel a, div.star-rating a {
-  display: block;
-  width: 16px;
-  height: 100%;
-  background-position: 0 0px;
-  border: 0
-}
-
-div.star-rating-on a {
-  background-position: 0 -32px !important
-}
-
-div.star-rating-hover a {
-  background-position: 0 -32px
-}
-
-/* Read Only CSS */
-div.star-rating-readonly a {
-  cursor: default !important
-}
-
-/* Partial Star CSS */
-div.star-rating {
-  background: transparent !important;
-  overflow: hidden !important
-}
-
-/* END jQuery.Rating Plugin CSS */
