+++ lib/generators/spree_reviews/install/install_generator.rb
@@ -4,8 +4,7 @@ module SpreeReviews
       class_option :auto_run_migrations, type: :boolean, default: false
 
       def add_javascripts
-        append_file 'vendor/assets/javascripts/spree/frontend/all.js', "//= require spree/frontend/spree_reviews\n"
-        append_file 'vendor/assets/javascripts/spree/backend/all.js', "//= require spree/backend/spree_reviews\n"
+        # none
       end
 
       def add_stylesheets
