+++ lib/spree_reviews.rb
@@ -2,7 +2,6 @@ require 'spree_core'
 require 'spree_reviews/engine'
 require 'spree_reviews/version'
 require 'spree_extension'
-require 'coffee_script'
 require 'sass/rails'
 
 module Spree
